Build linux-amd64 binary in installer-artifacts
This will be used by oc in the next release to allow extracting an
x86_64 installer from another architecture's payload.

Due to the size of the installer binary, building each in a separate
stage minimizes resource requirements when using imagebuilder.  This
does require a matching change to ocp-build-data to add a second builder
stage.