crio: fix bootstrap given new crio config behavior
now, it's not guaranteed that crio will print out every field in the config
this is good, it allows us to iterate on default values quicker. However, this means we can't use simple tools like sed to update the config
luckily, `crio config` is now primed to create new drop-in files to configure itself. Use this behavior to set the override

Signed-off-by: Peter Hunt <pehunt@redhat.com>