baremetal: stop provisioning services once control plane is deployed
Currently keepalived monitors only the kubernetes API before determining
whether to failover to the control plane. However, we also use the API
VIP to host provisioning services since 22c32f8, so we need to include
Ironic in the health checks.  However, we need to stop Ironic when we're
done to trigger failover.

This change makes Ironic shutdown once all control plane hosts are
"active" (i.e. deployed), and the introspection data is synced.