Bug 1956208: openstack: add validation for volume types
When an user wants to deploy with rootVolume, they need to specify
a volume type.
We already checked that the volume type was set but we didn't check that
the type was actually existing.

With this patch, we now check that the type exists.

If the Volume Type doesn't exist, it'll show this error message:

```
FATAL failed to fetch Metadata: failed to load asset "Install Config": compute[0].platform.openstack.rootVolume.type.type: Invalid value: "wrong": Volume Type either does not exist in this cloud, or is not available
```

Note: this patch also adds test coverage when the type is an empty
string, which wasn't done before.