platform: Disallow publish internal for non-cloud installations
The publish: internal strategy is only available in cloud platforms,
GCP, Azure, AWS, Alibaba and IBM cloud, and should not be allowed as a
strategy for the other platforms.

Added a validation check to check if the strategy is internal and
the platform used is one of the valid cloud platforms. Check for IBM
and Alibaba is not yet added due to them not being supported yet.