baremetal: Do not use ironic allocation
While testing secure boot scenario where only 1 master is using,
we found that the instance_info capabilities for secure_boot was set
in a different master.
This was happening because in all allocations we were considering
all nodes, since we need to match 1-1 the information we don't need
to create allocations for each node.