Bug 1965921 - [oVirt] High performance VMs shouldn't be created with Existing policy
RHV changed the behavior of creating High-Performance VMs.
In the past when creating a High-Performance VM it would be created with Existing auto pining policy, but now Existing is removed and shouldn't be used.

We need to update the terraform provider that is used within the Openshift installer to consume the changes.