vsphere: use fully qualified path for datacenter
This change ensures we are using the fully qualified path of the
vsphere datacenter everywhere where we interact with the vsphere API.
By doing so, we remove the errors when duplicate datacenters with the
same name exist in the environment. The change also modifies the
questionaire to set the fully qualified path for the datacenter in the
install-config. It is backwards compatible with prior install-configs in
that it will prepend a leading slash on datacenter names that do not
already have it.