Azure Stack: add machine configs
NOTE The need for these machine configs should be reassessed when
out-of-tree cloud providers are implemented.

On Azure Stack, the cloud provider configuration for the Kubelet
utilizes an endpoints file and an environment variable to set the path
to that file. cf:
https://kubernetes-sigs.github.io/cloud-provider-azure/install/configs/#azure-stack-configuration

This commit adds machine configs for masters and workers to write the
endpoints file to disk and set the environment variable on the kubelet
with a systemd drop in.

When changing endpoints, users will need to update the machine configs.