asset/cluster: use terraform stages with outputs fed between stages
Use the Stage interface to run the terraform applies and destroys.
The outputs from each stage are used as input into subsequent stages.
This allows stages to get information that would either not be available
at all or only available via a data source. For example, with GCP there
is a stage where the health check IDs are needed. There is no data source
available in the gcp terraform provider for accessing these IDs. However,
using the output from the stage where the health checks are created,
the stage is able to get the health check IDs as input instead.