terraform: means for establishing complex terraform stages
Add a Stage interface which defines a given terraform stage for a
given platform. For each stage, there is an option to include the
stage in the destroy when doing a bootstrap destroy. There is also
an option to provide a custom destroy function. For example, when
destroying the GCP bootstrap, there is a stage where the destroy
behavior is modifying the load balancers to remove the bootstrap
from the backend.