cmd: consider all stages when looking for ips for bootstrap gather
Include the IPs of the bootstrap instance and the control plane
instances in the output of the terraform stage where the instances
are created. This is used by the bootstrap gather as a consistent
way to grab the IPs as opposed to extracting the IPs from the
terraform state files.

When looking for the IPs of the bootstrap and control plane machines,
consider all of the terraform stages. For compat stages, there is
no change in behavior. For newer split stages, the gather looks through
the output files for the "bootstrap_ip" and "control_plane_ips"
outputs.