Remove unreachable special case
We had code to allow you to not specify an IPv6 machineNetwork value
on dual-stack AWS clusters, but this is a no-op since we no longer
allow you to configure dual-stack on AWS anyway. And if we support it
in the future, it will likely be after improvements to IPv6 support on
the AWS end, so this exception may no longer make sense at that point
anyway.