aws: move creation of bootstrap instance profile to cluster stage
When the bootstrap instance profile is created just before the bootstrap
instance is created, then the RunInstance call will return an invalid
value error for a little while. The terraform provider waits up to 2 minutes
for this to resolve, which is does in that time. However, the terraform
provider can, and frequently does, lose a response to a subsequent
RunInstance. When this happens the terraform provider does not re-attempt
the RunInstance request and instead sits idly waiting for the 2-minute
timeout to elapse.

To combat this, we move the creation of the bootstrap instance profile to
the cluster stage. This will create a delay between when the instance profile
is created and when the instance is created. This is due to the instance
profile being created early in the stage and there being a long delay while
waiting for other resources in the stage to be created.

https://bugzilla.redhat.com/show_bug.cgi?id=1975475