aws: block creation of bootstrap instance until ignition config is uploaded
Give the bootstrap instance a dependency on the creation of the s3 bucket
object holding the bootstrap's ignition config. The bootstrap instance
requires the ignition config in order to complete its creation. It takes
around 2 minutes to create the bucket object, and the instance creation has
a 2-minute timeout. Consequently, there are times when the bootstrap fails
to complete in time because the creation of the bucket object took too long.

https://bugzilla.redhat.com/show_bug.cgi?id=1975475