Azure Stack: add AAD Client Credentials to Cloud Provider Config
This is a temporary addition to add the client credentials to the cloud
provider config to support bootstrapping the kubelet with the legacy
cloud provider. Once the Azure out-of-tree provider has been implemented
we can utilize a merged cloud provider config similar to ARO so that the
client credentials are saved in a secret rather than in plaintext on the
nodes.

From my reading of the legacy provider in the kubelet, the merged config
is not supported when bootstrapping (hence the need for this commit).
The call for bootstrapping is made here:
https://github.com/openshift/kubernetes/blob/master/staging/src/k8s.io/legacy-cloud-providers/azure/azure.go#L360

But that code never calls getConfigFromSecret, which creates the merged
config:
https://github.com/openshift/kubernetes/blob/master/staging/src/k8s.io/legacy-cloud-providers/azure/azure_config.go#L66

Instead, getConfigFromSecret is called from Initialize:
https://github.com/openshift/kubernetes/blob/master/staging/src/k8s.io/legacy-cloud-providers/azure/azure.go#L675

which appears to only be called from the kube-controller-manager:
https://github.com/openshift/kubernetes/blob/master/cmd/kube-controller-manager/app/controllermanager.go#L601

This is not a problem for Public Azure because it uses managed identity,
which is not supported in Azure Stack.