aws: move elastic ip permissions to create networking category
Elastic IPs are only created when the installer provisions VPCs and
subnets. In the case where users install into an existing VPC, the
installer does not need the permissions to create and associate EIPs.
Therefore let's move these EIP permissions so that they are only checked
when we create networking resources.