Default to use-octavia for openstack cloud config
Use octavia if available, unless using kuryr.  There should be no
penalty for setting `use-octavia` in your cloud configuration even when
octavia is not available in the environment. Many people didn't realize
they needed to enable the octavia integration as a day 2 operation if
they wanted to use LB type services on an OpenStack platform. This is
now performed for them at install time and is less disruptive for the
cluster. Also, in the event the cloud platform gains the Octavia service
after the initial installation, then the openshift cluster should be
able to use it out of the box.