Add SRIOV Support for Kubevirt Provider
Currently, the kubevirt VirtualMachines that are created during the cluster provisioning, are hard-coded with Bridge interface binding, despite the actual type of the network that was configured in install-config.yaml.
This PR adds an optional new field to install-config.yaml, specifying the network type to be used when creating the kubevirt VirtualMachines. Example:
platform:
  kubevirt:
    apiVIP: <api_vip>
    ingressVIP: <ingress_vip>
    namespace: kubevirt-nested-cluster
    networkName: <NAD name>
    networkType: SRIOV          # <------- New Field
    storageClass: <storage class used>

Signed-off-by: orenc1 <ocohen@redhat.com>