Allow to set the "InfraID" with an environment variable
This change allows the user to set the "OPENSHIFT_INSTALL_INFRA_ID_OVERRIDE"
environment variable to modify the "InfraID".

The contents of the "OPENSHIFT_INSTALL_INFRA_ID_OVERRIDE" correspond to
the random part of the InfraID so it must be formed by 5 random lowercase
alphanumeric chars with no vowels, otherwise an error will be returned.

An usage example would be:

 OPENSHIFT_INSTALL_INFRA_ID_OVERRIDE="xxxxx" openshift-install create manifests --dir=.

This change is needed to implement external provider support in the
assisted-installer project.