Add azurestackprivate Terraform Provider
Adds support for an azurestack Terraform provider in order to upload a
local RHCOS VHD OS image to an Azure Stack storage blob. This provider
makes use of azure-vhd-utils to speed up download, because an RHCOS VHD
is a fixed 16GB size but most of that is zero-padding. The
azure-vhd-utils determines the amount of data that needs to be uploaded
and cuts out the empty ranges, which in effect results in a 2.5GB upload
rather than 16GB.

Unfortunately the azure-vhd-utils library is quite old and uses version
5 of the azure-sdk-for-go (currently we are on version 55). Version 5 of
  the sdk was created before semantic versioning was in place in the
repo, therefore there is no graceful way to import version 5 of the SDK
because other parts of the installer depend on later versions of the
SDK.

To work around this issue, I have incorporated the upload package from the
azure-vhd-utils directly into the terraform provider and updated it to
import the openshift fork of the azure-sdk-for go, pinned to version 5.
This technical debt would need to be incurred until the azure-vhd-utils
is updated to a newer version of the SDK, which may never happen.

The config.go is modeled after the azurestack provider, so the majority
of the original work is in the resource_vhd_blob.go and then some wiring
in azurestackprivate.go and provider.go.