baremetal: Ensure ipv6 bootstrap VM client-id is predictable
In #4052 we added interfaces to control the MAC addresses for the
NICs in the bootstrap VM, so that in environments where DHCP pools
are not allowed, a static reservation can be made.

Unfortunately that doesn't work in ipv6 environments currently,
because NetworkManager needs a specific configuration similar to
that applied via the MCO[1] for cluster hosts to ensure the
generated client-ID is derived from the MAC and predictable.

[1] https://github.com/openshift/machine-config-operator/blob/master/templates/common/on-prem/files/NetworkManager-onprem.conf.yaml