Azure: cloud provider config excludeMastersFromStandardLB -> false
This updates the Azure cloud provider config to set
excludeMastersFromStandardLB to false. The default value is true, which
has not been a problem because the functionality was broken. With the
move to out-of-tree providers the functionality has been restored, in
which case if a master node restarts the service controller will not add
it back to the load balancer backend pool.