Bug 1993364: openstack/destroy: fix Kuryr/BYON
When using byon with Kuryr the Router would be identified based
on the Primary Network used by Servers, which would be filtered
by the device-owner compute:nova. However, when using azs the
device-owner name can change. This commit fixes the issue by
identifying the router by looking for any tagged networks that
has a subnet connected to the router. This approach can't be
run before other clean ups because the CNO would always attempt to
re-connect the  service subnet into the router.

Co-Authored-By: Maysa Macedo <maysa.macedo95@gmail.com>
Co-Authored-By: Emilien Macchi <emilien@redhat.com>