Bug 1978213: openstack/quota: relax min ports
For Kuryr:
Reduce from 1493 to 1490:
* 6 ports taken by machines (and added later in the requirements)
* 2 for floating IPs
* 1 for router gateway
* 1 for bootstrap machine
= 10
1500 - 10 = 1490
So we have now 1490 as a new relaxed number which should satisfy the
documented quotas.

For non-Kuryr:
Reduce from 8 to 4:
* 1 taken by network gateway
* 2 taken by API & Ingress ports or FIPs
* 1 taken by bootstrap
-> this makes 4
Then the installer adds one port per machine, so 6 by default (3 masters
and 3 workers) which makes 10 total. The documented quota is 15 so have
some margin error.