AWS: Write user credentials to specified env location
If the user specified the AWS_SHARED_CREDENTIALS_FILE which is
either invalid or empty file, the credentials are collected by
the installer and are written to a default location (~/.aws/credentials)
when it can be written to the location specified in the above location.

Adding a small fix to write to the location specified in the location
specified by AWS_SHARED_CREDENTIALS_FILE if the file is either empty
or is invalid.