AWS: Stop searching other China regions for resources
For cluster destroy on AWS china regions, the cn-northwest-1
region is always searched for resources even if the cn-north-1
region is used for installation. This might be true for the
public regions where there are a few resources in the us-east-1
location but the china regions are separate partitions and are
not involved in the other region installation.

Removing the northwest region client from the destroy so it
will not be used to search if it is not used for installation.