Azure: Display proper error message on failure to delete
If there are other resources in the resource group that are in use
by resources in other groups then the deletion will be blocked.
The error message displayed is not user friendly and hence adding
a check to see if the error code returned by azure autorest is
409, which is ResourceGroupDeletionBlocked, and display a proper
message to the user.