machines: Extract root disk sizes to constants
I was looking at this issue:
https://github.com/openshift/openshift-docs/issues/35793
Which then led me to wonder exactly what the defaults are per
platform for disk sizes.

And then I stumbled into this 120 vs 128 thing that as best
I can tell from the git history is mostly an accident.  If someone
happens to know differently (e.g. can/should we unify on 120 or 128?)
then that'd be good.

In the meantime, an additional benefit here is that if one does
e.g. `git grep -i volume.*size` these constants will turn up.
It took me a little while to figure out how to backtrack from
the Terraform code to this.  (The lines here only contained
`Size` which is too generic to grep)