hack/verify-vendor.sh: clean module cache pre-test
go mod verify "checks that dependencies of the main
 module stored in the module cache have not been
modified since they were downloaded." As these tests
are running in a container, it the contents of that
cache seem arbitrary.

By adding go clean -modcache we can clear the cache
before testing to ensure a fresh start and that
go mod verify is not complaining about altering
contents that should be removed.

[0] https://golang.org/ref/mod#go-mod-verify