pkg/destroy/gcp: report cluster footprint in quota
When we destroy an entity in the public cloud relating to a cluster, we
now record the impact that the item had to the quota in the account that
the cluster was provisioned in. This will allow for downstream users of
the installer to reason about the footprint of clusters they run,
allowing for more automated reasoning about how many clusters of a type
can fit into an account.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>