baremetal: improve host role management during assets creation
The logic for generating baremetal manifests and Terraform variables now takes
into account the host role tag properly, if defined. This allows to define
master/worker hosts in any order. To remain backward compatible, if an host
does not have any role defined (or supported value), it's considered eligible
for the ControlPlane or Compute respectively, based on the configured
requirements. The validation rule has been enhanced to verify that the correct
number of hosts have been configured.