update-rhcos-bootimage: check that RHCOS metadata has necessary artifacts
It is possible to feed the `update-rhcos-bootimage.py` script a
`meta.json` input that does not contain all the boot image artifacts
required when doing a boot image bump. This changes the script to bail
out if the `images` dict is missing well known artifacts that are
expected when doing an x86_64 boot image bump.