Bug 2008119: force cvo to ignore installer-provided resources
When the installer provides manifests for resources that are included
as create-only resources in the release payload, there is a race
between whether cluster-bootstrap will apply the installer-provided
manifest or cluster-version-operato will apply the payload-provided
manifest. Whichever one is applied first will be the only one applied.
This is problematic since the expectation is that the installer-provided
manifests should always be applied.

To get around this, the installer now adds to the .spec.overrides field
of the ClusterVersion resource all of the resources for which the installer
is providing manifests. This tells the cluster-version-operator to not
apply any manifests for these resources. After the cluster-bootstrap
completes, the bootkube service will remove those resources from the
.spec.overrides field so that the cluster-version-operator can take over
management of those resources during upgrades.

The cvoignore asset has been added during the ignition configs
stage. This asset is responsible for creating two files for the
bootstrap node: cvo-overrides.yaml and original_cvo_overrides.patch.

The cvo-overrides.yaml file is a replacement of the one created by the
Common Manifests asset. The replacement adds to the .spec.overrides field
all of the resources included in the Common Manifests and Openshift Manifests
assets.

The original_cvo_overrides.patch file is a json patch that the bootkube
service applies to the ClusterVersion resource after cluster-bootstrap
completes. The patch restores any value that the user supplied for the
.spec.overrides field.