gather-bootstrap: execute dbginfo script to collect debugging info on IBM Z systems
This is to address https://bugzilla.redhat.com/show_bug.cgi?id=2012326. Today as part of the docs for IBM Z installation [1], we ask users to run a script to collect debug information[2] by doing an oc debug node. There should be no need for this as this should be part of the must-gather and the installer-gather command in case of bootstrapping failures.

[1] https://access.redhat.com/documentation/en-us/openshift_container_platform/4.8/html/installing/installing-with-rhel-kvm-on-ibm-z-and-linuxone
[2] https://github.com/ibm-s390-linux/s390-tools/blob/master/scripts/dbginfo.sh