validation: Enforce strict unmarshalling of config
In order to stop the user from adding unrecognized properties
in the install config, a validation check is added to force the
user to remove them. This is done by enforcing a strict
unmarshalling of the install config yaml file that will raise an
error when an unknown property is added.

This will force the installer to return an error for the first
occurence of an unknown property since the current library
returns only the first unknown field.