cmd/openshift-install/create: One shot console access
The console may become optional [1], so teach the installer to handle
its absence gracefully.

We've waited on the console since way back in ff53523b11 (add logs at
end of install for kubeadmin, consoleURL, 2018-12-06, #806).  Back
then, install-complete timing was much less organized, and since
e17ba3c571 (cmd: wait for the cluster to be initialized, 2019-01-25, #1132)
we've blocked on ClusterVersion going Available=True. So the current
dependency chain is:

1. Console route admission blocks console operator from going
   Available=True in its ClusterOperator.
2. Console ClusterOperator blocks cluster-version operator from
   going Available=True in ClusterVersion.
3. ClusterVersion blocks installer's waitForInitializedCluster.

So we no longer need to wait for the route to show up, and can fail
fast if we get a clear IsNotFound.  I'm keeping a bit of polling so we
don't fail an install on a temporary network hiccup.

We don't want to drop the console check entirely, because when it is
found, we want:

* To continue to log that access pathway on install-complete.
* To continue to append the router CA to the kubeconfig.

That latter point has been done since 40335776f6 (append router CA to
cluster CA in kubeconfig, 2019-02-12, #1242).  The motivation in that
commit message is not explicit, but the idea is to support folks who
naively run 'oc login' with the kubeadmin kubeconfig [2] (despite that
kubeconfig already having cluster-root access) when the console
route's cert's CA happens to be something that the user's local trust
store doesn't include by default.

[1]: https://github.com/openshift/enhancements/pull/922
[2]: https://github.com/openshift/installer/pull/1541#issuecomment-480896291